requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pandas==2.2.3
numpy==2.2.5
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.1.0